The code of uaf::Variant has been made more generic (by using macros) to minimize the amount of code needed.
Also, Variants now store native UAF objects for NodeIds, ExpandedNodeIds and QualifiedNames. These particular UAF types (uaf::NodeId, uaf::ExpandedNodeId and uaf::QualifiedName) store the namespace URI in addition to the namespace index. The SDK types on the other hand, only store the latter. This can lead to a loss of information. E.g. when a QualifiedName with a server URI is stored in a VariantVector, previously the server URI information was lost (because the internal UaVariant object wrapped by uaf::Variant is not able to store server URIs). As of now, this information is stored.