Tweak helper scripts to prepend uaf\lib directory to environment variables
Prepending the uaf\lib directory assures that the correct libraries are
found e.g. for executing the tests.

I had an entry in the PATH environment variable which pointed to an OPC
UA SDK 1.3.3 - that caused the error "ImportError: DLL load failed: The
specified procedure could not be found." when running the tests. Tests
on master require OPC UA SDK 1.4.x.